chore(server/conn): backport `max_pending_accept_reset_streams()` to builder
NB: this commit is based on the `0.14.x` release series.

the `server::conn::http2` submodule provides types that are
conditionally compiled when the `backports` feature is active, to
facilitate upgrading to the `1.x` release.

i've been working on upgrading the
[`linkerd2-proxy`](https://github.com/linkerd/linkerd2-proxy/) project
to use hyper 1.0, but encountered some gaps in `Builder<E>`'s interface
when setting the `backports` and `deprecated` features.

this adds a `max_pending_accept_reset_streams(..)` method
(_added in #3201_) to this builder, so that 0.14.x users relying on
this functionality can prepare to upgrade to hyper 1.x safely.

this is effectively a backport of #3507, which restored this interface
to the 1.0 release.

for more information, see:

* https://github.com/hyperium/hyper/pull/3201
* https://github.com/hyperium/hyper/pull/3507
* https://github.com/hyperium/hyper/issues/3461
* https://hyper.rs/guides/1/upgrading/

Signed-off-by: katelyn martin <me+cratelyn@katelyn.world>