chore(client/conn): `SendRequest` conversion glue
users of the `backports` feature that are attempting to migrate from
0.14 to 1.0 may need to convert from a legacy `SendRequest` (returned
by some of the connector builders) to an instance of the new protocol
specific `conn::http2::SendRequest<B>` or
`conn::http1::SendRequest<B>` senders.

this adds `From<T>` implementation to the type provided by the
`backports` feature, to help facilitate incremental migration.

Signed-off-by: katelyn martin <me+cratelyn@katelyn.world>