Allow recording the raw HTTP/1 headers sent and received
This allows feeding them into a WARC file
(https://en.wikipedia.org/wiki/WARC_(file_format)). The rest of the
request and response is already available as it's either the explicitly
set body or the received response body.

This includes the final `\r\n\r\n` between the headers and the response
(so that it can be distinguished from bare `\n\n`).

Needed for a similar request in reqwest:
https://github.com/seanmonstar/reqwest/issues/1229.