test(body): add proper cfgs to body unit tests (#3864)
Running `cargo test` fails to compile because tests rely on types which
are behind a #[cfg(…)] which is disabled by default.

Add a #[cfg(…)] directive to tests which rely on types which are also
behind a #[cfg(…)] directive, so that these tests run only if the types
on which they depend exist.